requirements.txt
absl-py
astor
cachetools
certifi
chardet
Django
django-cors-headers
djangorestframework
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
joblib
Keras
Keras-Applications
Keras-Preprocessing
Markdown
numpy
oauthlib
opencv-python
opt-einsum
pandas
Pillow
protobuf
pyasn1
pyasn1-modules
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
sqlparse
tensorboard
tensorflow
tensorflow-estimator
termcolor
urllib3
Werkzeug
wrapt
